+++ app/src/main/java/com/evertec/everteplacetopay/vo/json/input/Taxe.kt
@@ -2,6 +2,6 @@ package com.evertec.everteplacetopay.vo.json.processTransaction.input
 
 data class Taxe(
     val amount: Double,
-    val base: Int,
+    val base: Double,
     val kind: String
 )
