+++ app/src/main/java/com/evertec/everteplacetopay/vo/json/input/To.kt
@@ -2,5 +2,5 @@ package com.evertec.everteplacetopay.vo.json.input
 
 data class To(
     val currency: String,
-    val total: Int
+    val total: Double
 )
